test/web-platform-tests/to-upstream/dom/attributes-are-not-nodes.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Attributes are not Nodes and should not be accepted where they are</title>
<link rel=help href="https://dom.spec.whatwg.org/#dom-core-changes">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

test(() => {

  const attribute = document.createAttribute("newattribute");

  const notOnAttrs = ['nodeType', 'appendChild', 'normalize'];

  for (const prop of notOnAttrs) {
    assert_false(prop in attribute, `${prop} should not exist on attribute instances`);
    assert_false(prop in Attr.prototype, `${prop} should not exist on Attr.prototype`);
  }

}, 'Attrs do not have nodeType properties');

test(() => {

  const parent = document.createElement("p");

  const attribute = document.createAttribute("newattribute");
  assert_throws(new TypeError(), () => {
    parent.appendChild(attribute);
  });

}, "appendChild with an attribute should fail");

test(() => {

  const parent = document.createElement("p");
  parent.appendChild(document.createElement("span"));

  const attribute = document.createAttribute("newattribute");
  assert_throws(new TypeError(), () => {
    parent.replaceChild(attribute, parent.firstChild);
  });

}, "replaceChild with an attribute should fail");

test(() => {

  const parent = document.createElement("p");
  parent.appendChild(document.createElement("span"));

  const attribute = document.createAttribute("newattribute");
  assert_throws(new TypeError(), () => {
    parent.insertBefore(attribute, parent.firstChild);
  });

}, "insertBefore with an attribute should fail");

</script>
